chore: 🤖 remove console logs
