+++ src/components/AgentsUI/index.tsx
@@ -5,59 +5,13 @@ import {
 } from '@components/Eliza/api';
 import { useAuthStore } from '@fleek-platform/login-button';
 import ElizaIntegrationLayer from '@components/Eliza/ElizaIntegrationLayer.tsx';
-import { getCookie } from '@utils/cookies';
+import { getReferralId } from '@utils/promotekit';
 
-export interface AgentsUIIntegrationProps {
-  getReferralId?: () => string;
-}
-
-const getReferralIdFromCookie = () => {
-  console.log('[debug] eliza.astro: getReferralIdFromCookie: ', 1)
-
-  try {
-     const promotekit_referral = getCookie('promotekit_referral');
-
-    if (!promotekit_referral) throw Error('Promotekit referral cookie was not found!');
-
-    console.log('[debug] eliza.astro: getReferralIdFromCookie: ', promotekit_referral)
-
-    return promotekit_referral;
-  } catch (_err) {
-    console.warn(
-      `User session is not a Promotekit referral`
-    );
-
-    return '';
-  }
-}
-
-const getReferralId = () => {
-  console.log('[debug] eliza.astro: getReferralId: ', 1)
-
-  try {
-    console.log('[debug] eliza.astro: getReferralId : ', window?.promotekit_referral)
-
-    if (!window.promotekit_referral) throw Error('Promotekit referral not found in global window object!');
-
-    return window.promotekit_referral;
-  } catch (_err) {
-    console.warn(
-      `Promotekit referral is not available. Will check cookie`
-    );
-    return getReferralIdFromCookie();
-  }
-};
-
-export const AgentsUIIntegration: React.FC<AgentsUIIntegrationProps> = (
-  props,
-) => {
+export const AgentsUIIntegration = () => {
   const { triggerLoginModal, accessToken, isLoggingIn, isLoggedIn, projectId } =
     useAuthStore();
   const login = () =>
     typeof triggerLoginModal === 'function' && triggerLoginModal(true);
-
-  console.log(`[debug] AgentsUI: typeof getReferralId = ${typeof getReferralId}`);
-  console.log(`[debug] AgentsUI: on init: getReferralId(): ${getReferralId && getReferralId()}`);
   
   return (
     <ElizaIntegrationLayer
@@ -74,8 +28,8 @@ export const AgentsUIIntegration: React.FC<AgentsUIIntegrationProps> = (
   );
 };
 
-const AgentsUI: React.FC<AgentsUIIntegrationProps> = (props) => (
-  <AgentsUIIntegration {...props} />
+const AgentsUI = () => (
+  <AgentsUIIntegration />
 );
 
 // to be used in Astro
