+++ src/components/Eliza/components/SubscriptionModal.tsx
@@ -42,19 +42,12 @@ export const SubscriptionModal: React.FC<SubscriptionModalProps> = ({
 
     const referralId = getReferralId ? getReferralId() : '';
 
-    console.log(`[debug] window.promotekit_referral = ${window.promotekit_referral}`)
-
-    console.log(`[debug] SubscriptionModal: on handleSubmitClick: getReferralId: ${typeof getReferralId == 'function' && getReferralId()}`);
-
-
     const subscriptionCreationResponse = await createSubscription(
       activeProjectId,
       productId,
       accessToken,
       referralId,
     );
-
-    console.log('SubscriptionModal - referralId', referralId);
 
     if (
       !subscriptionCreationResponse ||
